Use float type math

+++ algorithms/rateServoFullNonlinear/rateServoFullNonlinearAlgorithm.cpp
@@ -80,7 +80,7 @@ CmdTorqueBodyMsgF32Payload RateServoFullNonlinearAlgorithm::update(const uint64_
     if (this->Ki > 0.0F) { /* check if integral feedback is turned on  */
         this->z += omega_BBast_B * dt;
         for (Eigen::Index i = 0; i < 3; i++) {
-            const float intLimCheck = fabs(this->z[i]);
+            const float intLimCheck = fabsf(this->z[i]);
             if (intLimCheck > this->integralLimit) {
                 this->z[i] *= this->integralLimit / intLimCheck;
             }
